Create the Adapters Namespace
